refactor(mcp): split monolithic index.mts into modular file structure
