+++ @xen-orchestra/mcp/src/helpers/tool-error.mts
@@ -1,0 +1,3 @@
+export function formatToolError(error: unknown): string {
+  return error instanceof Error ? error.message : 'Unknown error occurred'
+}
