+++ @xen-orchestra/mcp/src/tools/query/get-infrastructure-summary.mts
@@ -1,0 +1,81 @@
+import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
+import type { XoClient } from '../../xo-client.mjs'
+import { formatToolError } from '../../helpers/tool-error.mjs'
+
+export function registerGetInfrastructureSummary(server: McpServer, getClient: () => XoClient): void {
+  server.registerTool(
+    'get_infrastructure_summary',
+    {
+      title: 'Get Infrastructure Summary',
+      description: 'Get a high-level summary of the entire XO infrastructure (pools, hosts, VMs counts)',
+      inputSchema: {},
+    },
+    async () => {
+      try {
+        const client = getClient()
+
+        const [pools, hosts, vms] = await Promise.all([
+          client.listPools('id,name_label'),
+          client.listHosts({ fields: 'id,name_label,power_state' }),
+          client.listVms({ fields: 'id,power_state' }),
+        ])
+
+        const runningVms = vms.filter(vm => vm.power_state === 'Running').length
+        const haltedVms = vms.filter(vm => vm.power_state === 'Halted').length
+
+        const summary = {
+          pools: {
+            total: pools.length,
+            names: pools.map(p => p.name_label),
+          },
+          hosts: {
+            total: hosts.length,
+          },
+          vms: {
+            total: vms.length,
+            running: runningVms,
+            halted: haltedVms,
+            other: vms.length - runningVms - haltedVms,
+          },
+        }
+
+        return {
+          content: [{ type: 'text', text: JSON.stringify(summary, null, 2) }],
+        }
+      } catch (error) {
+        return {
+          content: [
+            {
+              type: 'text',
+              text: `Failed to get infrastructure summary: ${formatToolError(error)}`,
+            },
+          ],
+          isError: true,
+        }
+      }
+    }
+  )
+
+  server.prompt(
+    'infrastructure-overview',
+    'Generate a natural language overview of the XO infrastructure',
+    async () => {
+      return {
+        messages: [
+          {
+            role: 'user',
+            content: {
+              type: 'text',
+              text: `Please use the get_infrastructure_summary tool to fetch the current state of the Xen Orchestra infrastructure, then provide a clear, human-readable summary of:
+1. How many pools are configured and their names
+2. Total number of hosts
+3. VM statistics (total, running, halted)
+
+Present this information in a friendly, conversational way.`,
+            },
+          },
+        ],
+      }
+    }
+  )
+}
