+++ @xen-orchestra/mcp/src/tools/query/list-pools.mts
@@ -1,0 +1,36 @@
+import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
+import { z } from 'zod'
+import type { XoClient } from '../../xo-client.mjs'
+import { formatToolError } from '../../helpers/tool-error.mjs'
+
+export function registerListPools(server: McpServer, getClient: () => XoClient): void {
+  server.registerTool(
+    'list_pools',
+    {
+      title: 'List Pools',
+      description: 'List all pools in Xen Orchestra with their basic information',
+      inputSchema: {
+        fields: z
+          .string()
+          .optional()
+          .describe(
+            'Comma-separated fields to return (default: id,name_label,name_description,auto_poweron,HA_enabled)'
+          ),
+      },
+    },
+    async ({ fields }) => {
+      try {
+        const client = getClient()
+        const pools = await client.listPools(fields)
+        return {
+          content: [{ type: 'text', text: JSON.stringify(pools, null, 2) }],
+        }
+      } catch (error) {
+        return {
+          content: [{ type: 'text', text: `Failed to list pools: ${formatToolError(error)}` }],
+          isError: true,
+        }
+      }
+    }
+  )
+}
